Now a regular user is also created on fresh install
