+++ .env
@@ -11,3 +11,5 @@ SECRET_KEY="+xp-e^wzkp$83qnh-aamck(ng6zmqg42@h2k!7m$y%3r$j*#kg"
 DJANGO_SUPERUSER_USERNAME=root
 DJANGO_SUPERUSER_PASSWORD=root
 DJANGO_SUPERUSER_EMAIL=root@django.com
+REGULAR_USER_USERNAME=user
+REGULAR_USER_PASSWORD=user
