+++ docker-compose.yml
@@ -9,13 +9,15 @@ services:
     command: >
       sh -c "python manage.py wait_for_database &&
              python manage.py migrate &&
-             python manage.py create-superuser &&
+             python manage.py create-users &&
              python manage.py runserver 0.0.0.0:8000"
     environment:
       - SECRET_KEY=${SECRET_KEY}
       - DJANGO_SUPERUSER_USERNAME=${DJANGO_SUPERUSER_USERNAME}
       - DJANGO_SUPERUSER_PASSWORD=${DJANGO_SUPERUSER_PASSWORD}
       - DJANGO_SUPERUSER_EMAIL=${DJANGO_SUPERUSER_EMAIL}
+      - REGULAR_USER_USERNAME=${REGULAR_USER_USERNAME}
+      - REGULAR_USER_PASSWORD=${REGULAR_USER_PASSWORD}
       - DB_HOST=${DB_HOST}
       - DB_NAME=${POSTGRES_DB}
       - DB_USER=${POSTGRES_USER}
